Replaced code with regex
Based on feedback from Paul, I ended up rewriting the original property
implementation for ReleaseEntry.PackageName.  It uses a regex to parse
out the package name from a release entry.  The regex could quite
possible be used for finding the release type and also version as I have
groups in the regex for such.