Unescape entries when we initially build our full package
Unfortunately, when we removed escaping in 1.7.0, while we removed it everywhere
for our own code, we forgot about the input - when people build their NuGet
packages, NuGet will still do the escaping nonsense. We need to un-escape our
input file, which sucks out loud because we can't use 7-Zip there, but :cry:

This bug is really bad because it breaks npm scoped packages, one of which is
in the box in every electron-forge app.