Expose an $(SquirrelToolsPath) property for nuget consumers
When using the nuget package, it's often convenient to
be able to execute it in some targets. However, since the nuget
install path depends on the version of the package, it's typically
harder than necessary to figure out the path to tools.

This adds a .props file that is automatically imported by nuget
when installed in a project, therefore making it trivial to invoke
the tools by just using the new $(SquirrelToolsPath) MSBuild property
from anywhere in the project or its imported targets.

Since it's in a .props, it means that even downstream packages that
add a dependency on this package can leverage the tool too without
having to know what version of the Squirrel.Windows package is in use.

Fixes #1364