Improved doc for build based on #1492
Gives more details for which option needs to be checked in the visual installer, since it's now standardized for the last 2 versions of Visual Studio.
I attempted to explain it in a way that won't require updates often. Not sure how to give more details about the version of VC++ without having something that would need frequent updates.