Added CLI flag to onlyUpdate shortcuts
By default, running withcreateShortcut will create shortcuts in each specified location. If the user has deleted one of these shortcuts, it should optionally not be recreated. `CreateShortcutsForExecutable` has an option to "onlyUpdate" shortcuts (i.e. skipping if the shortcut does not already exist, this change esposes this option to the CLI as a flag.